% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cases_current_strategy}
\alias{cases_current_strategy}
\title{Time series of hospitalisations over time with current strategy.}
\format{
A data frame with 306 rows and 2 variables:
\describe{
\item{dates}{\code{string}, date}
\item{hospitalisations}{\code{int}, hopitalisation cases}
}
}
\usage{
cases_current_strategy
}
\description{
A dataset containing the dates and hospitalisation numbers for the current strategy.
}
\keyword{datasets}
