% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cases_current_strategy_flattened}
\alias{cases_current_strategy_flattened}
\title{Time series of hospitalisations over time with current strategy including infection-curve flattening.}
\format{A data frame with 306 rows and 2 variables:
\describe{
\item{dates}{\code{string}, date}
\item{hospitalisations}{\code{int}, hopitalisation cases}
}}
\usage{
cases_current_strategy_flattened
}
\description{
A dataset containing the dates and hospitalisation numbers for the current strategy with infection-curve flattening.
}
\keyword{datasets}
