% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_simr-gamma.R
\name{plot.Covidsimr}
\alias{plot.Covidsimr}
\title{Plot output of COVID-19 simulation}
\usage{
\method{plot}{Covidsimr}(x, ...)
}
\arguments{
\item{inputs}{PLACEHOLDER}

\item{outputs_cum}{PLACEHOLDER}
}
\description{
Plot output of COVID-19 simulation
}
