% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_simr-gamma.R
\name{plot_cum_died_inscap}
\alias{plot_cum_died_inscap}
\title{Plot cumulative total of patients who died as a result of insufficent capacity over simulation period}
\usage{
plot_cum_died_inscap(outputs_cum, display = TRUE)
}
\arguments{
\item{display}{}
}
\value{

}
\description{
Plot cumulative total of patients who died as a result of insufficent capacity over simulation period
}
