% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_simr-gamma.R
\name{plot_cum_total_nonadmitted_surv}
\alias{plot_cum_total_nonadmitted_surv}
\title{Plot cumulative total of patients who could not be admitted because of capacity constraints but did
surivive, over simulation period}
\usage{
plot_cum_total_nonadmitted_surv(outputs_cum, display = TRUE)
}
\arguments{
\item{display}{}
}
\value{

}
\description{
Plot cumulative total of patients who could not be admitted because of capacity constraints but did
surivive, over simulation period
}
