% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_simr-gamma.R
\name{plot_daily_bed_occupancy}
\alias{plot_daily_bed_occupancy}
\title{Plot daily bed occupancy over simulation period}
\usage{
plot_daily_bed_occupancy(outputs, cap, display = TRUE)
}
\arguments{
\item{display}{}
}
\value{

}
\description{
Plot daily bed occupancy over simulation period
}
