% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_simr-gamma.R
\name{plot_daily_deaths}
\alias{plot_daily_deaths}
\title{Plot daily deaths result from insufficient capacity over simulation period}
\usage{
plot_daily_deaths(outputs, display = TRUE)
}
\arguments{
\item{display}{}
}
\value{

}
\description{
Plot daily deaths result from insufficient capacity over simulation period
}
