% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_simr-gamma.R
\name{plot_daily_hospitalisations}
\alias{plot_daily_hospitalisations}
\title{Plot daily hospitalisations (including resampling and tolerance to give confidence ranges)}
\usage{
plot_daily_hospitalisations(inputs, outputs, uncert, display = TRUE)
}
\arguments{
\item{display}{}
}
\value{

}
\description{
Plot daily hospitalisations (including resampling and tolerance to give confidence ranges)
}
