% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_simr-gamma.R
\name{print.Covidsimr}
\alias{print.Covidsimr}
\title{Tidy print \code{Covidsimr} result.}
\usage{
\method{print}{Covidsimr}(x, ...)
}
\arguments{
\item{x}{a \code{covidsimR} object.}

\item{...}{any other parameters}
}
\value{
a tidy print of the return data.
}
\description{
Tidy print \code{Covidsimr} result.
}
