% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_simr-gamma.R
\name{simfn}
\alias{simfn}
\title{COVID-19 ICU bed occupancy simulation}
\usage{
simfn(rep, dates, metric, value, galpha, gbeta, cap, pfat, uncert)
}
\arguments{
\item{rep}{(\code{postive integer}), number of repetitions in simulation.}

\item{dates}{(\code{vec, Date}), dates in siulation.}

\item{metric}{PLACEHOLDER}

\item{value}{PLACEHOLDER}

\item{galpha}{(\code{positive numeric}), median length of stay for admitted covid19 pateints, to be estimated by user from data or other source}

\item{gbeta}{(\code{positive numeric}), 0.95 quantile length of stay for admitted covid19 patients, to be estimated by user from data or other source}

\item{cap}{(\code{positive integer}), maximum number of patients who can be concurrently admitted to the hospital/unit}

\item{pfat}{(\code{numeric} in range 0 (no chance of death)`  to 1 (certain death)), probability that a patient who arrives but is rejected/cannot be admitted (because there are no beds free) dies}

\item{uncert}{(\code{logical})}
}
\value{

}
\description{
COVID-19 ICU bed occupancy simulation
}
