% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_simr-gamma.R
\name{plot_cum_admissions}
\alias{plot_cum_admissions}
\title{Plot cumulative total of admitted patients over simulation period}
\usage{
plot_cum_admissions(outputs_cum, display = TRUE)
}
\arguments{
\item{display}{}
}
\value{

}
\description{
Plot cumulative total of admitted patients over simulation period
}
