% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_simr-gamma.R
\name{plot_daily_bed_occupancy}
\alias{plot_daily_bed_occupancy}
\title{Plot daily bed occupancy over simulation period}
\usage{
plot_daily_bed_occupancy(outputs, cap, display = TRUE)
}
\arguments{
\item{outputs}{\code{data.frame}, output data from simulation.}

\item{cap}{\code{int}, the bed capacity.}

\item{display}{\code{log}, defaults to \code{TRUE}. Should the plot display or not.}
}
\value{
Time-siries plot of daily bed occupancy. Invisibly returns the plot object.
}
\description{
Plot daily bed occupancy over simulation period
}
