% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_simr-gamma.R
\name{plot_daily_deaths}
\alias{plot_daily_deaths}
\title{Plot daily deaths result from insufficient capacity over simulation period}
\usage{
plot_daily_deaths(outputs, display = TRUE)
}
\arguments{
\item{outputs}{\code{data.frame}, output data from simulation.}

\item{display}{\code{log}, defaults to \code{TRUE}. Should the plot display or not.}
}
\value{
Time-series plot of daily deaths from insufficient capacity. Invisibly returns the plot object.
}
\description{
Plot daily deaths result from insufficient capacity over simulation period
}
