% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_simr-gamma.R
\name{plot_daily_hospitalisations}
\alias{plot_daily_hospitalisations}
\title{Plot daily hospitalisations (including resampling and tolerance to give confidence ranges)}
\usage{
plot_daily_hospitalisations(inputs, outputs, uncert, display = TRUE)
}
\arguments{
\item{inputs}{PLACEHOLDER}

\item{outputs}{\code{data.frame}, output data from simulation.}

\item{uncert}{\code{log}, does the simulation include uncertainty.}

\item{display}{\code{log}, defaults to \code{TRUE}. Should the plot display or not.}
}
\value{
Time series plot of daily hospitalistions. Invisibly returns the plot object.
}
\description{
Plot daily hospitalisations (including resampling and tolerance to give confidence ranges)
}
